requirements.txt
pillow
requests
beautifulsoup4
lxml